created abstract class to reduce redundant code (ta feedback)
