+++ src/main/entities/Question.java
@@ -3,19 +3,13 @@ package main.entities;
 import java.util.ArrayList;
 
 /** Entity to represent a question. */
-public class Question {
+public class Question extends AbstractQuestionLike{
     /** Question body text. This is the question the players will be asked. */
     private final String questionText;
     /** Correct answer to the question. */
     private final String correctAnswer;
     /** ArrayList of incorrect answers. */
     private final ArrayList<String> incorrectAnswers;
-    /** Difficulty of the question. */
-    private final String difficulty;
-    /** Category of the question. */
-    private final String category;
-    /** Type of the question (multiple or boolean). */
-    private final String type;
 
     /**
      * Create a mew question with the given parameters.
@@ -29,12 +23,10 @@ public class Question {
      */
     public Question(String questionText, String correctAnswer,
                     ArrayList<String> incorrectAnswers, String difficulty, String category, String type){
+        super(category, difficulty, type);
         this.questionText = questionText;
         this.correctAnswer = correctAnswer;
         this.incorrectAnswers = incorrectAnswers;
-        this.difficulty = difficulty;
-        this.category = category;
-        this.type = type;
     }
 
     /**
@@ -62,33 +54,6 @@ public class Question {
      */
     public ArrayList<String> getIncorrectAnswers(){
         return incorrectAnswers;
-    }
-
-    /**
-     * Retrieve the difficulty of the question.
-     *
-     * @return the difficulty of the question.
-     */
-    public String getDifficulty(){
-        return difficulty;
-    }
-
-    /**
-     * Retrieve the category of the question.
-     *
-     * @return the category of the question.
-     */
-    public String getCategory(){
-        return category;
-    }
-
-    /**
-     * Retrieve the question type.
-     *
-     * @return the question type.
-     */
-    public String getType(){
-        return type;
     }
 
 }
